fix: fix bug with crash when click "send message", fix bug with location(create new chat), fix bug bottom bar

+++ app/src/main/java/io/moonshard/moonshard/db/ChooseChatRepository.kt
@@ -3,8 +3,8 @@ package io.moonshard.moonshard.db
 import io.moonshard.moonshard.models.api.Category
 
 object ChooseChatRepository {
-    var lat:Float?=null
-    var lng:Float?=null
+    var lat:Double?=null
+    var lng:Double?=null
     var address:String = ""
     var time:String = ""
     var category:Category? = null
